+++ changelog.md
@@ -13,6 +13,7 @@ This version is not released yet and is under active development.
 - \[mpm\] Run bar plugin unittests in their independent, non-parrallel step.
 - \[mpm\] Skip testing on intermediate Python versions to speed up CI. Only the oldest and latest supported.
 - \[mpm\] Skip configuration-related tests while we investigate test isolation.
+- \[mpm\] Fix fetching of full local copy of cask tap in tests to allow for checkout of past formula.
 - \[mpm\] Replace unmaintained `bump2version` by `bump-my-version`.
 
 ## {gh}`5.13.1 (2023-05-04) <compare/v5.13.0...v5.13.1>`